refactor(libolm-compat): Move the libolm unpickling into a separate module
